Scraper/ARRIVEALIVE_Scraper/scraper.py
"""
Scraper for the Arrive Alive website.
"""
import random
from typing import List
import requests
from bs4 import BeautifulSoup
from ..base_scraper import BaseScraper, AccidentRecord


class ArriveAliveScraper(BaseScraper):
    """
    Scraper for the Arrive Alive website.
    """
    def __init__(self):
        super().__init__("ARRIVEALIVE")
        self.base_url = "https://www.arrivealive.co.za/road-safety-statistics"
        self.raw_data = None

    def fetch_data(self) -> None:
        """
        Fetch data from the Arrive Alive website.

        In a real implementation, this would make HTTP requests to the website
        and download the data. For this boilerplate, we'll simulate the process.
        """
        try:
            # In a real implementation, this would be:
            # response = requests.get(self.base_url)
            # self.raw_data = response.text

            # For now, we'll simulate the data
            self.raw_data = """
            <html>
                <body>
                    <table>
                        <tr>
                            <th>Region</th>
                            <th>Accidents</th>
                            <th>Year</th>
                        </tr>
                        <tr>
                            <td>Western Cape</td>
                            <td>1200</td>
                            <td>2020</td>
                        </tr>
                        <tr>
                            <td>Gauteng</td>
                            <td>2500</td>
                            <td>2020</td>
                        </tr>
                        <tr>
                            <td>KwaZulu-Natal</td>
                            <td>1800</td>
                            <td>2020</td>
                        </tr>
                    </table>
                </body>
            </html>
            """
            print(f"[ARRIVEALIVE] Successfully fetched data from: {self.base_url}")
            print(f"[ARRIVEALIVE] Exact scraping location: {self.base_url}")
        except Exception as e:
            print(f"[ARRIVEALIVE] Error fetching data from {self.base_url}: {e}")
            # In a real implementation, you might want to retry or handle the error differently

    def parse_data(self) -> None:
        """
        Parse the fetched data and convert it to AccidentRecord objects.

        In a real implementation, this would parse the HTML or other format
        and extract the relevant information. For this boilerplate, we'll
        simulate the process.
        """
        if not self.raw_data:
            raise ValueError("No data to parse. Make sure to fetch data first.")

        try:
            # In a real implementation, this would parse the HTML:
            # soup = BeautifulSoup(self.raw_data, 'html.parser')
            # table = soup.find('table')
            # rows = table.find_all('tr')[1:]  # Skip header row
            # 
            # for row in rows:
            #     cells = row.find_all('td')
            #     region = cells[0].text.strip()
            #     accident_count = int(cells[1].text.strip())
            #     year = int(cells[2].text.strip())
            #     
            #     record = AccidentRecord(region, accident_count, year, self.source_name)
            #     self.records.append(record)

            # For now, we'll simulate the parsing with the data from our simulated HTML
            regions = ["Western Cape", "Gauteng", "KwaZulu-Natal"]
            for region in regions:
                # Simulate data for 2018-2022
                for year in range(2018, 2023):
                    accident_count = random.randint(1000, 3000)
                    record = AccidentRecord(region, accident_count, year, self.source_name)
                    self.records.append(record)

            print(f"[ARRIVEALIVE] Successfully parsed {len(self.records)} records")
        except Exception as e:
            print(f"[ARRIVEALIVE] Error parsing data: {e}")
            # In a real implementation, you might want to handle the error differently
